KOGITO-1609: Moved CoverageReport dock in BC module
